Split into KeepHelper
